requirements.txt
appdirs==1.4.3
bcrypt==3.1.3
cffi==1.9.1
Django==1.9.11
django-countries==4.5
django-easycart==0.4.0
django-extensions==1.7.5
django-guardian==1.4.1
django-userena==2.0.1
docutils==0.13.1
easy-thumbnails==2.3
html2text==2014.12.29
olefile==0.44
packaging==16.8
Pillow==4.0.0
pycparser==2.17
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
